[CID-B-12]: editing entities and Dto´s

+++ ProjectWorkspace/CID/CID-Core/src/main/java/com/cid/beans/SalesDto.java
@@ -1,19 +1,17 @@
 package com.cid.beans;
 
-import java.util.Date;
-
 import lombok.Data;
 
 @Data
 public class SalesDto {
-	private Long id;	
+	private long id;	
 	private String date;
 	private int week;
 	private String month;
-	private int eurPerHour;
-	private int paxPerHour;
-	private int eurPerPax;
-	private int callsPerHour;	
+	private double eurPerHour;
+	private double paxPerHour;
+	private double eurPerPax;
+	private double callsPerHour;	
 	
 	private UserDto user;
 }
